Początek wyszukiwarki produktów i filtrowania

+++ queries/karty_del_query.php
@@ -1,7 +1,7 @@
 <?php
     session_start();
-    //if(isset($_SESSION['admin'])){
+    if(isset($_SESSION['admin'])){
         require_once('db.php');
         $query = mysqli_query($conn, "DELETE FROM cards WHERE name='{$_SESSION['index']}'");
-    //}
+    }
 ?>
